добавил во вьюхи методы установки обработчиков событий, отрефакторил utils
