+++ src/utils/common.js
@@ -16,11 +16,4 @@ const shuffleArray = (arr) => {
   return shuffledArray;
 };
 
-const humanizeFilmDuration = (runtime) => {
-  const hours = Math.floor(runtime / 60);
-  const minutes = runtime % 60;
-
-  return `${hours}h ${minutes}m`;
-};
-
-export {getRandomInteger, shuffleArray, humanizeFilmDuration};
+export {getRandomInteger, shuffleArray};
