+++ src/view/btn-more-view.js
@@ -6,4 +6,18 @@ export default class BtnMoreView extends AbstractView {
   get template() {
     return createBtnMoreTemplate();
   }
+
+  setClickHandler = (cb) => {
+    this._callback.click = cb;
+    this.element.addEventListener('click', this.#clickHandler);
+  };
+
+  removeClickHandler = () => {
+    this.element.removeEventListener('click', this.#clickHandler);
+  };
+
+  #clickHandler = (evt) => {
+    evt.preventDefault();
+    this._callback.click();
+  };
 }
